Add card management in Create Beneficiary modal stepper
