--- Sig.App.Frontend/src/views/beneficiary/_AssignCardForm.vue
@@ -1,16 +1,0 @@
-<i18n>
-  {
-    "en": {
-      
-    },
-    "fr": {
-      
-    }
-  }
-  </i18n>
-
-<template>
-  <p>TODO :: Assignation d'une carte</p>
-</template>
-
-<script setup></script>
